Add useMapLayer hook
This hook useful for adding a layer to the map that is mutable. It will
be used by a few other features later so it's worth extracting to a
general purpose hook now.
